+++ CHANGELOG.md
@@ -35,6 +35,11 @@ Section Order:
 ### Security
 -->
 
+### Changed
+
+- Using stopword lists downloaded from https://github.com/stopwords-iso to improve
+  search results
+
 ## \[2.1.1\] - 2024-04-28
 
 ### Added